Better tiling for GEMMs.
This change does the following:
 1. Choose GEMM tile sizes that fit into the L1 or L2 data cache,
 2. Flip the order in which the GEMM tiles are computed (loop over `m`, then `n`).

With the first change, the tile size `mr`x`nc` is chosen such that the `mr`x`k` and `nc`x`k` rows/columns of the inputs/weights are still in the L1 cache (or L2 if `k` is too large to fit in L1 even with `nc`=`nr`) at the end of the computation.

With the second change, once a tile of size `mr`x`nc` has been computed, the next tile can re-use the `nc`x`k` weight values already in the L1/L2 cache and only needs to read the new `mr`x`k` input values.

Since `nc` is usually much larger than `mr`, this tile ordering significantly improves cache re-use (provided the tile data fits in the L1/L2 cache).

PiperOrigin-RevId: 715811605