Fix msan for gemm compiled outputs
msan is a compiler instrumentation tool, but because the gemm compiler directly generates assembly, the necessary annotations are missing.

This change manually instruments the results of gemms to be initialized, by inserting a tail call to a function with the same signature as a gemm, which unpoisons the output of the gemm.

To facilitate this, we need to save and restore a few more registers to preserve the arguments for the tail call.

PiperOrigin-RevId: 724247504