Use the new `pthreadpool_parallelize_[23]d_tile_2d_dynamic` strategies in the GEMM-based ops.
This results in a 0%-10% speedup across a number of GEMM-heavy benchmarks for one or more cores on a number of different architectures.

The performance gains are due mainly to:
 * On one or more threads, fewer individual calls to the operator function resulting in less function call overheads (most noticeable for small GEMMs),
 * On multiple threads, fewer overheads within `pthreadpool` since elements are processed in chunks, not individually (most noticeable for small GEMMs),

Note that these improvements are **not due to better load balancing**, since the granularity of the parallel computations is actually larger. What these parallelization strategies bring to the table is:
 * Fewer overheads due to larger granularity,
 * Equivalent load balancing _despite_ the larger granularity.

Due to this, the results are better _even for models that anti-scale_ with the number of threads, i.e. increasing the number of threads makes performance worse, but with this CL, the performance is _less worse_.

PiperOrigin-RevId: 724355125