Fix hanging benchmarks on platforms without cpuinfo
If we didn't have cpuinfo, we tried to flush caches of 128MB in size. This means that for small benchmarks that take ~1us, we would spend orders of magnitude more time flushing the cache than benchmarking, and because we pause timing while flushing the cache, the benchmark framework tries to run the benchmark for hours to get enough samples.

PiperOrigin-RevId: 727053860