Add an outer loop over channels to depthwise convolution kernels (~25% speedup for large channels)
By doing this, the input and weights can be kept in L1 cache. It would be better to have this loop outside the microkernel (so we could parallelize over this loop too), but indirection buffers make this difficult.

This adds a little overhead for the shallow (small `channels`) cases, but it's negligible in the cases I've measured.

PiperOrigin-RevId: 730587324