Add outer loop over channels to f16 and qs8 dwconv microkernels, ~50% speedup in some cases
By doing this, the input and weights can be kept in L1 cache. It would be better to have this loop outside the microkernel (so we could parallelize over this loop too), but indirection buffers make this difficult.

This adds a little overhead for the shallow (small channels) cases, but it's negligible in the cases I've measured.

This is similar to #7869 (same thing for f32)

PiperOrigin-RevId: 730657776