Rewrite avgpool microkernels
- Remove multipass
- Use xnn simd wrappers.
- Generate code via templates.
- A bunch of general cleanups and bug fixes, removing multipass makes a lot of code *much* simpler and easier to find bugs in.
- In particular, I'm pretty sure there were horrible latent bugs regarding mixed up usage of scaleminmax vs. minmax params.

This is a bit slower (5-15%) than the previous implementation in most cases, because it needs to load from `prev_output` in addition to the inputs. I think this is a better tradeoff, because the old version required two separate codepaths, one for one pass, and another for multiple passes. I could fork these cases too, but I don't think it makes sense for an operator that is rarely significant for overall performance.

If we really care about the performance here, we can now trivially generate avx2 and avx512 versions of these kernels, which should much more than recover the loss in performance. We could also use the same optimization from #7876, which enabled dwconv unipass kernels to outperform the multipass kernels.

PiperOrigin-RevId: 731173104