Rewrite maxpool kernels
- Use templates + simd wrappers to generate kernels
- Don't read past the end of the indirection buffer

The kernels are very similar to what we had previously (except f16c, which now uses a signcomplement integer approach), it just removes the primary tile/incremental tile as an implementation detail that is made visible to callers due to reading past the end of the indirection buffer. By cleaning up that overreading (these are scalar reads outside the loop over channels, performance impact is negligible), it makes calling these kernels much simpler.

PiperOrigin-RevId: 731917763