Attempt to speed up gemm benchmark builds taking ~1 hour on github
Currently, there is a sequence of lambdas and functions that are generated for every benchmark x kernel combination, resulting in 7 symbols per benchmark. This is apparently hitting a bad algorithm in some linkers.

This change eliminates the wrapper lambda by using a template argument instead (doesn't change function pointer type, no need for wrapper lambda). The number of symbols for each benchmark is reduced from 7 to 4.

PiperOrigin-RevId: 732220150