Fix RVV vrnd losing infinity
The RVV vrnd ukernels work by converting to int and back. This loses the "special" float values (inf, NaN).

This change tries to at least preserve inf, matching NaN is a little trickier and also not quite as bad to lose (but we should find a way to do so).

Unfortunately, I don't see any RVV intrinsics that would implement rounding without converting to int.

PiperOrigin-RevId: 734774226