Remove channel_subtile and channel_round from dwconv_config
These were never set to anything other than channel_tile and `1`, respectively. This is a ton of complexity that we just weren't using at all, and we're unlikely to want to use in the future.

There are also likely bugs in this code if we were to actually try to use this complexity. Note that the microkernel tester did not actually use `channel_subtile` or `channel_round`! It just passed `channel_tile` for all 3 parameters... The benchmarks and packing test were similar too.

PiperOrigin-RevId: 736061436