Initialize memory in pack functions that kernels expect to be initialized
This changes packing to not assume that the output has been initialized by a prior memset call. This pattern is a dangerous source of bugs. In addition to that, it requires memsetting the whole pack buffer, which may be less efficient than just initializing memory appropriately in the packing functions. The scalar kernels changed here are actually a bit faster than they were before this CL.

The case where this is questionable in terms of performance is batch matrix multiply, where packing may be called many times on a buffer that is initialized once. IMO, relying on previous initialization is a risky decision that makes a lot of sketchy assumptions. It would be worth a small performance penalty to avoid that.

There are two kinds of uninitialized memory addressed by this change:
- Uninitialized weights that need to be a "zero" value for the gemm kernel. Note these small holes don't exist in the pack microkernels, it's only the reference functions and scalar microkernels that have a lot of small holes.
- `extra_bytes`, which is an opaque blob to the pack function. This is more problematic. As a temporary workaround, this change memsets the packed weights buffer as it did prior to this change, but only when `extra_bytes > 0` in the packed buffer.

PiperOrigin-RevId: 736603771