Initialize weights in pack functions that kernels expect to be initialized
This changes packing to not assume that the output has been initialized by a prior memset call for weights. This pattern is a dangerous source of bugs.

This is a partial roll forward of a previous change that caused breakages. This change only modifies the packing functions to fully initialize the weights.

PiperOrigin-RevId: 736685560