Disable c2 gemm kernels on x86
These kernels produce numerically inconsistent results depending on whether the CPU we are running on supports AVX512 or not. This causes brittle tests to fail inconsistently.

Temporarily, we are going to disable these kernels, but we are going to re-enable them in the future because they offer significant performance improvements in some cases.

PiperOrigin-RevId: 738547744