Produce better distribution of random floats for testing
A uniform distribution of floats between -FLT_MAX and FLT_MAX is a pretty bad distribution for testing arithmetic. It tends to produce a lot of infinity/NaN, and relatively few numbers are "near" 0.

To work around this, generate random floats by reinterpreting random integers as floats instead, and fix up the denormals/NaNs.

This also disables testing the rvv vrnd kernels (#8087).

PiperOrigin-RevId: 738865874