Avoid globally initializing pack buffers
This change uses the fact that pack functions can now be relied on to fully initialize the biases and weights. It still does not assume that "extra bytes" are initialized, and so we still memset the buffers in that case.

There is also a mysterious case where if we do not memset memory that is allocated by a weights cache, we will crash.

PiperOrigin-RevId: 739807980