Better SIMD wrapper implementation of `roundf` for `sse2`, which lacks a dedicated `round` instruction.
The problem with casting from `float` to `int32` and back is that this clobbers both non-finite values (`+/-Inf` and `NaN`), as well as values above/below `2^32`.

However, `roundf(x) == x` for any `fabs(x) >= 8388608.0f` (which is `2^32`) since there is no room for any fractional bits in the 23-bit mantissa.

We can therefore filter for `fabs(x) < 8388608.0f`, which will conveniently also return `false` for non-finite values, and then only convert to `int32` and back where this is `true`.

PiperOrigin-RevId: 742206010