Remove scaled dot product attention operator
This is unused, and marked experimental. Going forward, we would prefer that this operator be expressed as a graph or more basic operators, as the bench/models/ benchmarks do.

This already lacks features that newer models want, and I think that we can't reasonably support all the variations that exist with this approach. Building a subgraph is the way to go.

PiperOrigin-RevId: 743420802