Split CMake into subdirectories
This should reduce the chances of merge conflicts, and just generally is IMO cleaner and easier to work with. For example, right now if we were to add a target `foo` to both test and bench folders, I'm not sure what would happen, because both would create a binary `foo` in the build/local folder. This change eliminates that concern, because we'd now get a `foo` in both the `bench/` and `test/` subfolders.

I verified that ctest sees 475 tests before and after this change.

PiperOrigin-RevId: 745308956