Remove gemminc kernels
These are unused currently, not hooked up to any configs.

We are very interested in tiling the K dimension of gemms, which these kernels would be helpful with. However, it's not clear that these are on the critical path to doing this:
- This approach doesn't generalize beyond fp32 accumulators.
- There are some issues with these kernels that are likely undiscovered because this code is unused (e.g. min/max is applied at every tile of K, could use multiple different params to work around of course, but I think this just hasn't been thought about yet, and there are probably many other things like that).
- Packing that exists currently is not compatible with existing packing approaches, and would need quite a bit of work to get it up to speed.

If we do end up wanting these kernels, we can always revert this change. But as it stands, these kernels have been unused for years, and are unlikely to be used in the future.

PiperOrigin-RevId: 746084904