Move loops from around reduce microkernel testers to inside the `Test` function
This hopefully will avoid logs growing to be too big to be shown in bazel --test_output=errors tests on GitHub actions, which makes it impossible to see why a test failed.

This will also greatly increase the coverage of values we test (because the random number generator is seeded inside the Test function).

PiperOrigin-RevId: 748479110