Use simpler clamping in f32 -> qx8 conversions
The clamp we need already happens due to a saturating conversion to uint8 or int8 at the end of the kernel. This extra clamp exists to avoid dumb overflow behavior from `_mm*_cvtps_epi32`. For this clamp, we can do something simpler. Right now, we're basically double clamping exactly at the max, which is confusing, and I investigated it as a source of numerical inconsistency.

PiperOrigin-RevId: 758102002