Add missing `xnn_pf16_gemm_minmax_ukernel_1x32c2__neonsme2` microkernel which wraps KleidiAI's `kai_run_matmul_clamp_f16_f16_f16p2vlx2b_1x16vl_sme2_dot`, along with associated tests.
This CL also fixes several packing issues with `sme2`-enabled `fully-connected`, `dynamic-fully-connected` and `batch-matrix-multiply` exposed by the recent refactoring and extension of the corresponding subgraph unit tests.

In many cases, the `k_stride` parameter to the `pack_weights_and_biases` functions was being misinterpreted as the stride of the packed weight _outputs_, instead of the stride of the unpacked _inputs_.

As a bonus, this also adds a mechanism to name `xnnpack::Buffer` objects to make them easier to debug.

PiperOrigin-RevId: 758170249