Combine contiguous and discontiguous reduce configs
This fixes a terrible latent bug, another example of where the union strategy of defining operators goes haywire: the reduce and softmax configs both assign multiple configs. Before this CL, they just happened to not stomp on each other because the discontiguous config provided the right padding, but that was pure luck. This CL broke that, it took me a while to figure out why an rmax kernel was calling a vmul kernel...

PiperOrigin-RevId: 759642187