Replace `T params[restrict XNN_MIN_ELEMENTS(1)]` with `T* restrict params`
This was a 100% find-replace in the templates (.c.in) and then re-running all generator scripts.

I want to make this change because this syntax has given me a lot of headaches in the past, and it's giving me a headache now. I'm trying to reconcile the build flags between our builds, so external PR authors don't have to play guess and check to get their builds passing. I can either find the right combination of flags that makes this syntax work, or just remove it.

Benchmarking is unnecessary because the compiled binaries are identical.

PiperOrigin-RevId: 762098581