Find and fuse sequences of unary elementwise quantized ops into a single LUT op.
This enables sequences of operations like `softsign(x) = 1/(1 + abs(x))` or `softplus(x) = log(1 + exp(x))` to be optimized to a single LUT operation.

LUTs are not necessarily the fastest way to implement such operations, but they are much faster than our default behavior.

Sequences of operations will be fused as long as all ops in the sequence are pure, unary (or binary with a constant broadcasted operand) and elementwise, and the input and output of the sequence are quantized (but intermediate values may not be).

PiperOrigin-RevId: 762137227