Work around weird model issue in quantized unary op LUT fusion
This model has a value that is both static (with data) *and* an external input. I think this is probably a bug in the upstream code, but, it's easy enough for us to work around it.

The only bug fix in this change is the change to `is_broadcasted_static`. All the rest of the changes are just cleaning that I did for debugging, but I think they are good changes to keep.

PiperOrigin-RevId: 762505022