Be even more defensive about weird graphs in unary quantized fusion
There are graphs that give us tensors with static data that are also used as outputs. Fusing such values would result in a dead node.

This also fixes a bug that wasn't actually hit by anything yet: the operand of binary elementwise ops treated as unary (because the other operand is a constant) was backwards. However, this return value is only used if the binary op is the first in a fusion chain, which doesn't affect any existing case or test.

PiperOrigin-RevId: 762585395