Sort the `subgraph` nodes topologically before optimizing.
The sorting is stable and has a worst-case cost of O(`n`*`m`) where `n` is the number of nodes and `m` is the number of out-of-order nodes.

The sorting should not affect current models, but will make adding future `subgraph`
optimizations easier since many parts of the code quietly assume the `subgraph` nodes will always be topologically sorted.

PiperOrigin-RevId: 763442698