Fix undefined behavior in memcpy calls in softmax-nc.c
Due to a mistake in the C specification, memcpy cannot be passed NULL
pointers, even when the size is zero. This means code handling spans in
C needs to special case the empty slice everywhere.

Add a check to accommodate this in memcpy. This fixes a UBSan failure in
Chromium.

PiperOrigin-RevId: 767415600