Remove XNN_VALUE_FLAG_PERSISTENT
The only use of this has now been removed, and this flag is something we would rather not have if we don't need it, so let's remove it to avoid growing any new dependencies on it. This is not used by any of the major XNNPACK dependencies I'm aware of.

PiperOrigin-RevId: 767854924