Use XNN_FLAG_SLOW_CONSISTENT_ARITHMETIC to enable fast but numerically inconsistent kernels, or to enable slow `sse2fma` kernels to emulate faster architectures.
Before this change, this flag is effectively the default and only option. We've currently disabled a number of kernels that are faster, but cause results to be inconsistent.

After this change, those kernels are enabled, and the previous behavior is now opt-in via the flag. The new behavior is not identical to before: it now uses the faster kernels on new hardware, and to avoid the numerical inconsistency issue, we *emulate* the new hardware when it is not available. This means that when using the flag, the hardware that is penalized with lower performance (relative to before this CL) is *mostly* (but not entirely) old (pre-FMA) hardware, instead of new hardware.

PiperOrigin-RevId: 769281366