Set `_POSIX_C_SOURCE=199309L` for `bazel` builds.
This comes from cleaning up the includes in `subgraph.h`, which was unnecessarily pulling in `cache.h`, which itself was using `mutex.h`, which was setting the above.

Since this option affects a lot of system header files and needs to be set before any are included, it's best to set it as a compiler flag for _all_ sources.

Note that this is only needed for the GitHub `bazel` workflows. `cmake` is unaffected, i.e. the correct default is set implicitly.

PiperOrigin-RevId: 773586185