Add a more general experimental `xnn_scheduler_interface`
This replaces the previously added experimental `xnn_create_runtime_slinky` with a more general function. It happens to only support slinky at the moment, but perhaps we could attempt to support this with pthreadpool in the future if needed.

This scheduler interface is very minimal, and mirrors that found in other libraries we are likely to integrate with.

PiperOrigin-RevId: 778660251