Fix bug in `bias` initialization in `TestStaticB`.
The `bias` tensor was effectively never being set in this test. Setting it also requires setting the correct `bias_quantization`.

Testing with a bias tensor is still disabled for the `FullyConnectedF32QC[48]W` tests, since they use a `float`-valued bias which is then scaled by the channelwise factors of the weights, which differs significantly from the other GEMM kernels and thus has to be tested differently (this is, incidentally, covered by the operator tests).

PiperOrigin-RevId: 781994205