Add a hand-written assembly microkernel for `qd8_f32_qc4w_gemm_minmax` on `aarch32` and `neonmlal` based on the `qd8_f32_qc8w_gemm_minmax` kernel with `mr=4` and `nr=8`.
This is just a first tentative implementation focused mostly on correctness, not efficiency. The entire plan is to:
 1. Figure out how to implement this correctly (this CL),
 2. Automate generating such kernels via the GEMM compiler,
 3. Choose the best kernel for different `aarch32` variants.

The new `4x8` kernel is, however, already more than twice as fast as the current production `qd8_f32_qc4w_gemm_minmax_ukernel_6x16__neon_mlal_lane` kernel on a Pixel Watch 2.

PiperOrigin-RevId: 784599771