Add `xnn_threadpool` to enable sharing thread pools across multiple runtimes.
If an application creates multiple runtimes that cooperate on the same task, the runtimes should be able to steal work from each other to avoid waiting for loops to complete.

However, this is not possible before this change, because `xnn_scheduler` is used to make a new thread pool object for each runtime that is created via `xnn_create_runtime_with_scheduler`. While this object is inexpensive to construct, the objects do not steal work from each other.

This change solves that problem by adding an `xnn_create_threadpool` API, which creates an `xnn_threadpool` object that can be shared across multiple `xnn_create_runtime_with_threadpool` calls.

PiperOrigin-RevId: 791424869