Use a more generous tile size for the `unary-elementwise` and `binary-elementwise` op parallelization.
The L1 data cache size is used as a proxy for what could be considered a useful chunk of data on the underlying architecture. If the L1 data cache size is unknown, a default maximum chunk size of 32k is used.

This change provides better parallel efficiency in cases where thread contention is high, e.g. many unary/binary ops spread over a large number of threads.

PiperOrigin-RevId: 796868856