Create benchmarking buffers after reshape has been called.
Reshape overwrites/updates the shape of the output buffers, we shouldn't assume the output shape is correct until after reshape has been called.

This also required changing the method of getting the value size. `xnn_subgraph_value_get_size` doesn't (and can't) work because the reshaped size only exists in the runtime, not the subgraph.

PiperOrigin-RevId: 810743054