Allow kernel selection to use a transposed or not-transposed kernel if the transpose is a no-op.
If m = 1, or k = tile_k, then the transpose is a no-op, which means we don't need to limit the kernel selection to use a transpose_a kernel or not.

This means that we can use a non-SME kernel for m = 1, even if we thought we were going to use SME before we knew the shape of A.

This is less of an edge case than it seems, because m = 1 is common for batch size 1 use cases, and it's also a trivial optimization to implement, so there's no real downside.

PiperOrigin-RevId: 822789890