Remove `product` reduction type
We used to use this to support computing the normalization factor for means (the product of the extents of a value), but now we use `get_tensor_shape`'s reshaping functionality for that instead.

Since it's unoptimized/not well supported and probably has a lot of gotchas, I think we should just remove it.

PiperOrigin-RevId: 824797295