If B is not constant, reorder the loop over n outside the loop over m
And schedule pack_b to be computed outside the loop over m.

This enables dots with dynamic B to fully parallelize, with packing computed between the two loops.

This required implementing a bit of new logic in schedule(). This CL also contains a bunch of naming cleanup there. The change I ended up making is small, but I was experimenting a lot with this code.

PiperOrigin-RevId: 825902199