Remove slicing and expansion of quantization parameters based on k-dimensions.
This fixes a shape mismatch in the Convolution2DQC8 test which is now fully passes --runs_per_test=100. This is also needed for the Convolution2DQD8F32QC8W (there are some other issues with it though). I ran other tests and I don't think anything is broken by this.

PiperOrigin-RevId: 827620502