Fix `transpose` and `interleave` kernels generating incorrectly aligned loads
Even though we use `memcpy`, if the pointer is an int, the compiler will assume it is aligned to 4 bytes. However, this is not true in our case, because it's really an array of 4 chars being transposed at a time (in some cases).

We don't actually care about the type, we were just using that to communicate the size of an element. This change fixes that to be more explicit, and also allows instantiating `transpose` with a runtime size instead of a compile-time size (not currently used).

PiperOrigin-RevId: 828523286