Remove `xnn_define_all_input_shapes_static`.
This can be done manually by adding the `XNN_VALUE_FLAG_SHAPE_IS_STATIC` flag
value when creating the input values.

This means that the shapes must be set when creating the graph and ensures that
the person writing that code knows that these values are expected to be static.

PiperOrigin-RevId: 829459866