Add `dot_flag::consistent_arithmetic`
Before this change, we assumed kernels with tile_k = 1 were consistent. This was problematic in a few ways:
- It ignored fma vs. non-fma kernels
- For bf16 (and probably fp16 in the future), almost all kernels are tile_k = 2. So `YNN_FLAG_CONSISTENT_ARITHMETIC` left us with only the basic architecture independent kernel.

This change adds a flag `dot_flag::consistent_arithmetic`, where each type of kernel has a subset of kernels tagged with this flag. Currently, we are using a risky hack: on x86, we assume we can always use fma kernels. I hope this is a temporary situation and we can think of a better way.

Probably the most important thing this change does is add a test that kernels claiming to be numerically consistent, are.

A super interesting finding from working on this change is that bf16 doesn't care whether fma is used or not. I'm still not 100% sure about this, but I feel confident in the new test that checks this. That test immediately catches inconsistency between fp32 fma and non-fma kernels, so if there is a difference, it is some very special edge case.

Another "interesting" finding from this change is that as usual, Intel finds a way to be a headache. This time, it's because `dpbf16ps` does the reduction in a slightly different order than the "natural" choice: odd elements first, then even. This means that kernels using this instruction *can* be consistent with those that don't, but only if those other kernels swap the reduction order of the even and odd elements. We can do this easily for our tile_k = 2 kernels (and I did that in this change), but this means our default kernel is not numerically consistent with the rest.

PiperOrigin-RevId: 829566810