Don't lie about the supported architectures for the `get_dot_kernel` benchmarks
I wanted to benchmark a worst case for this function, i.e. all possible architectures are available.

However, this doesn't work for SME, because those kernels have an architecture defined block/tile shape, which is 0 if it isn't supported.

This was a latent bug, uncovered by an unrelated change, due to lack of type safety in the `get_dot_kernel` function. A more type safe way to handle this would be helpful.

PiperOrigin-RevId: 829627705