Use unaligned loads for B of x86 dot fp32 tile_k = 1 kernels
And use masked loads for the tail case of N.

If the data is actually aligned, this has no performance penalty on Intel Nehalem (2008) or AMD Bulldozer (2011) or later, and this will allow dots to run without packing.

This change only affects AVX or later kernels. While there is no performance impact of an unaligned load, there is an impact for pre-VEX encoding, because pre-VEX arithmetic ops can only use aligned memory operands, not unaligned. We also don't have a convenient way to do masked loads for pre-AVX, so rather than figure that out, we can just not bother.

In an abundance of caution, let's only do this for fp32 tile_k = 1 dots, because we probably can't run dots packless-ly for tile_k > 1 anyways.

PiperOrigin-RevId: 830534650