Enable dot subgraphs to use unpacked dot kernels without packing B
This is a major optimization in the case that B is dynamic, and A has only a few blocks of m. In this case, packing B is a waste of time, because we're only going to read B relatively infrequently anyways, and packing B costs a read itself, so in the worst case, this makes a dot cost ~2x as much.

PiperOrigin-RevId: 830749329