Remove tail case from avx512f dot tile_k = 1 kernels
It seems that masked loads are ~free on avx512, which means that it's pointless to have a tail case. This simplifies the code, and it also makes dots with `n % block_n != 0` faster, because we might be able to handle the dot with one instance of a block, instead of multiple instances of the tail, which is just one tile.

PiperOrigin-RevId: 831704537