Don't return `qdu8` gemm configs when the hardware doesn't support them.
`xnn_init_.*_gemm_config()` functions are expected to return NULL if the
hardware doesn't support the configuration. This is currently not the case for
the `xnn_init_qdu8_*_gemm_config()` functions, which silently return an
uninitialized structure even if the hardware doesn't support them.

We check against the `gemm_config::arch` field, because it is always set by
these functions when suitable hardware is found.

PiperOrigin-RevId: 832325627