Change int4 kernels from tile_n = 2 to tile_k = 2
This is more like what we will want to do in practice for optimized sub-byte datatype kernels.

The current motivation for this change is that we want to clean up the way strides are handled in `run_dot`, in order to make it easier to make `run_dot` handle packing of A and/or B dynamically. Pulling on that thread ended up here.

This is also an important change because the `pack` helper for dots currently is not able to handle 4-bit (or any sub-byte type) inputs. If the caller wanted to do a dot where B is 4-bit and transposed, we couldn't do that before this change.

PiperOrigin-RevId: 834773396