Add a new fingerprinting method for XNNPack weight cache.
Warning: this new fingerprinting method is currently experimental.

Every operation in XNNPack that uses the weight cache should implement a
corresponding fingerprinting function. The goal of that fingerprint is to check
whether the pre-computation (packing) that takes place for this op has changed
and if a previous weight cache can be reused directly or not.

The expected way for this to work is that a fingerprint function would assume a
predetermined set of constant inputs passed to the create function and would
then hash the resulting cache buffer.

The `xnn_weights_cache_look_up_key` has a new field that holds the fingerprint
identifier. It is set by kernel setup functions when they try to do cache
look-ups.

PiperOrigin-RevId: 834858245