Add test coverage of YNN_FLAG_CONSISTENT_ARITHMETIC
This doesn't test for consistency (which would be difficult to do), but we should test the correctness of the dot subgraph code with this flag set. Currently, there are few if any branches in the code based on this flag, but that will be changing soon.

This uncovered a bug: we didn't mark any arm kernels as consistent.

PiperOrigin-RevId: 835645726