Avoid interleave kernels in the tile_k = 1 case
When tile_k is 1, and we don't want to transpose the input, we currently use an interleave kernel, which in the case of interleaving one row, we just call memcpy.

This changes the logic to use a transpose kernel that basically does the same thing, but is both more clear what the meaning of the algorithm is, and also should be a little faster because it avoids an extra call via `interleave` in a loop that just wants to call `memcpy`, which is exactly what `transpose` does.

I think this is worth doing for clarity, and to avoid spurious appearances of `interleave` in profiling data.

This also changes `transpose` based on memcpy to take a parameter of bytes, not bits, since this function can't handle non-byte aligned element sizes.

PiperOrigin-RevId: 836246439