Remove `build_config.h` header
- Move definitions of `YNN_ARCH_X86`/`YNN_ARCH_ARM` to be defined in build rules instead of a header
- Move `YNN_HAVE_FLOAT16` logic to `half.h`.
- Move `YNN_ALLOCATION_ALIGNMENT` to `base.h`.

The main motivation here is that it's a little sketchy to need to include this header. We already have a mix of these two behaviors, the more specific architecture flags, e.g. `YNN_ARCH_X86_AVX2` are already defined this way, so this is making these flags more consistent.

I think we also have more reliable signals to determine this in the bazel BUILD configuration, or at least those signals are consistent with how we decide to use architecture specific libraries or features in the build.

PiperOrigin-RevId: 836426957