Use average pooling instead of convolution as a "dummy stencil"
These memory planner tests don't really test anything about convolution, they just need an op that incrementally reduces the size of the output due to being a stencil without padding. Average pooling is equivalent to convolution in that way.

This fixes an issue where convolution is complicated and messy, e.g. if ARM SME is available, we get inlining packing that changes the size of the workspace. Rather than account for that complexity in this test, let's just use average pooling instead.

PiperOrigin-RevId: 839201342