Optimize reduce kernel test
The main optimization is to create one large buffer for all of the test sizes we intend to test, and then just crop that buffer, significantly reducing the amount of random data we need to generate.

This generates test data with realistic crops too, so we don't need to explicitly test padding.

This also optimizes the reference implementation to avoid using Tensor::operator() as much.

Overall, this is a 3-4x speedup of the test, which is pretty significant when running on an emulator.

PiperOrigin-RevId: 839561665