Move bias add to after the dot operation.
Prior functionality added the bias prior to the dot operation by initializing the accumulator along with quantization offsets. This assumes that the bias has the same scale as the inputs.

New functionality adds the bias after the dot operation and uses the scale attached to the bias node (not the scale of the inputs). By moving this out, we are able to fuse it with other post-dot scaling operations.

PiperOrigin-RevId: 839642272