Fix out of bounds memory access when transposed A kernels have tile_m != 1
This was quite tricky to fix because there are really two bugs here: the out of bounds read, and writing the output in the wrong order, such that clamped rows could overwrite the valid rows.

Most of the changes here are just plumbing tile_m, which is currently not exposed.

PiperOrigin-RevId: 841809733