Use a more intuitive formulation of `make_stencil_dim`
At first I was surprised how this worked before, but it does make sense, because in this test, we only use the strides of the stenciled buffer, which are the same in these two formulations, only the extents change. So we are technically addressing this buffer out of bounds currently, but the actual addresses we compute are not changing before vs. after this change, so no actual memory is addressed out of bounds.

PiperOrigin-RevId: 842970311